+++ ccViewer/ccviewer.cpp
@@ -740,7 +740,8 @@ void ccViewer::toggleStereoMode(bool state)
 	if (isActive)
 	{
 		m_glWindow->disableStereoMode();
-		if (m_glWindow->getStereoParams().glassType == ccGLWindow::StereoParams::NVIDIA_VISION)
+		if (	m_glWindow->getStereoParams().glassType == ccGLWindow::StereoParams::NVIDIA_VISION
+			||	m_glWindow->getStereoParams().glassType == ccGLWindow::StereoParams::GENERIC_STEREO_DISPLAY)
 		{
 			//disable full screen
 			ui.actionFullScreen->setChecked(false);
@@ -780,7 +781,8 @@ void ccViewer::toggleStereoMode(bool state)
 			reflectPerspectiveState();
 		}
 
-		if (params.glassType == ccGLWindow::StereoParams::NVIDIA_VISION)
+		if (	params.glassType == ccGLWindow::StereoParams::NVIDIA_VISION
+			||	params.glassType == ccGLWindow::StereoParams::GENERIC_STEREO_DISPLAY)
 		{
 			//force full screen
 			ui.actionFullScreen->setChecked(true);
@@ -802,7 +804,10 @@ void ccViewer::toggleFullScreen(bool state)
 {
 	if (m_glWindow)
 	{
-		if (m_glWindow->stereoModeIsEnabled() && m_glWindow->getStereoParams().glassType == ccGLWindow::StereoParams::NVIDIA_VISION)
+		if (	m_glWindow->stereoModeIsEnabled()
+			&&	(	m_glWindow->getStereoParams().glassType == ccGLWindow::StereoParams::NVIDIA_VISION
+				||	m_glWindow->getStereoParams().glassType == ccGLWindow::StereoParams::GENERIC_STEREO_DISPLAY)
+			)
 		{
 			//auto disable stereo mode as NVidia Vision only works in full screen mode!
 			ui.actionEnableStereo->setChecked(false);
@@ -818,7 +823,12 @@ void ccViewer::onExclusiveFullScreenToggled(bool state)
 	ui.actionFullScreen->setChecked(m_glWindow ? m_glWindow->exclusiveFullScreen() : false);
 	ui.actionFullScreen->blockSignals(false);
 
-	if (!state && m_glWindow && m_glWindow->stereoModeIsEnabled() && m_glWindow->getStereoParams().glassType == ccGLWindow::StereoParams::NVIDIA_VISION)
+	if (	!state
+		&&	m_glWindow
+		&&	m_glWindow->stereoModeIsEnabled()
+		&&	(	m_glWindow->getStereoParams().glassType == ccGLWindow::StereoParams::NVIDIA_VISION
+			||	m_glWindow->getStereoParams().glassType == ccGLWindow::StereoParams::GENERIC_STEREO_DISPLAY)
+		)
 	{
 		//auto disable stereo mode as NVidia Vision only works in full screen mode!
 		ui.actionEnableStereo->setChecked(false);